backend/commons/gatling/gatling-context/src/test/java/com/kraken/runtime/context/gatling/environment/checker/ContainerNamesCheckerTest.java
package com.kraken.runtime.context.gatling.environment.checker;

import com.google.common.collect.ImmutableMap;
import com.kraken.runtime.entity.task.TaskType;
import com.kraken.tests.utils.TestUtils;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.test.context.ContextConfiguration;
import org.springframework.test.context.junit4.SpringRunner;

import static com.kraken.tools.environment.KrakenEnvironmentKeys.*;
import static org.assertj.core.api.Assertions.assertThat;

@RunWith(SpringRunner.class)
@ContextConfiguration(classes = {ContainerNamesChecker.class})
public class ContainerNamesCheckerTest {

  @Autowired
  ContainerNamesChecker checker;

  @Test
  public void shouldTest() {
    assertThat(checker.test(TaskType.GATLING_RUN)).isTrue();
    assertThat(checker.test(TaskType.GATLING_DEBUG)).isTrue();
    assertThat(checker.test(TaskType.GATLING_RECORD)).isTrue();
  }

  @Test(expected = NullPointerException.class)
  public void shouldFailCheck() {
    checker.accept(ImmutableMap.of());
  }

  @Test
  public void shouldSucceed() {
    final var env = ImmutableMap.<String, String>builder()
        .put(KRAKEN_GATLING_CONTAINER_NAME.name(), "value")
        .put(KRAKEN_GATLING_CONTAINER_LABEL.name(), "value")
        .put(KRAKEN_GATLING_SIDEKICK_NAME.name(), "value")
        .put(KRAKEN_GATLING_SIDEKICK_LABEL.name(), "value")
        .build();
    checker.accept(env);
  }

  @Test
  public void shouldTestUtils() {
    TestUtils.shouldPassNPE(ContainerNamesChecker.class);
  }
}
